Move Extract and WrapConst to separate files
