+++ core-v1/src/main/scala/algebra/WindowComparable.scala
@@ -58,40 +58,3 @@ object WindowComparable {
 
   implicit def identity[OP[_]]: WindowComparable[Id, OP] = anyIdentity.asInstanceOf[WindowComparable[Id, OP]]
 }
-
-trait WrapConst[F[_]] {
-
-  def wrapConst[A](value: A): F[A]
-}
-
-object WrapConst {
-
-  @inline final def apply[F[_] : WrapConst]: WrapConst[F] = implicitly
-
-  implicit val identity: WrapConst[Id] = new WrapConst[Lambda[a => a]] {
-    override def wrapConst[A](value: A): A = value
-  }
-
-  implicit val justified: WrapConst[Justified] = new WrapConst[Justified] {
-    override def wrapConst[A](value: A): Justified[A] = Justified.byConst(value)
-  }
-}
-
-// TODO: Replace with cats Monoid or ZIO Prelude?
-trait Extract[F[_]] {
-
-  def extract[A](fa: F[A]): A
-}
-
-object Extract {
-
-  @inline final def apply[F[_] : Extract]: Extract[F] = implicitly
-
-  implicit val identity: Extract[Id] = new Extract[Lambda[a => a]] {
-    override def extract[A](fa: A): A = fa
-  }
-
-  implicit val justified: Extract[Justified] = new Extract[Justified] {
-    override def extract[A](fa: Justified[A]): A = fa.value
-  }
-}
